+++ src/Bull/Core/Hardware/Win32/JoystickImpl.hpp
@@ -48,7 +48,7 @@ namespace Bull
              * \return Return the position of the axis
              *
              */
-            static float getAxisPosition(JoystickAxis axis, Uint8 joystick);
+            static unsigned int getAxisPosition(JoystickAxis axis, Uint8 joystick);
 
         private:
